organization(renamed components) added folders and made naming consistent
