+++ client/components/filterHomes/home-search.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { HomeSearchCard } from './index';
+import { HomeSearchCard } from '../index';
 /**
  * COMPONENT
  */
