+++ client/components/login-form.js
@@ -1,39 +1,49 @@
-import React from 'react'
-import {connect} from 'react-redux'
-import PropTypes from 'prop-types'
-import {auth} from '../store'
-
+import React from 'react';
+import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
+import { auth } from '../store';
 
 /**
  * COMPONENT
  */
-const AuthForm = (props) => {
-  const {name, displayName, handleSubmit, error} = props
+const AuthForm = props => {
+  const { name, displayName, handleSubmit, error } = props;
 
   return (
     <div>
       <form onSubmit={handleSubmit} name={name}>
         <div className="row">
           <div className="input-field col s10 m5">
-            <label htmlFor="email"><small>Email</small></label>
+            <label htmlFor="email">
+              <small>Email</small>
+            </label>
             <input name="email" type="text" />
           </div>
         </div>
         <div className="row">
           <div className="input-field col s10 m5">
-            <label htmlFor="password"><small>Password</small></label>
+            <label htmlFor="password">
+              <small>Password</small>
+            </label>
             <input name="password" type="password" />
           </div>
         </div>
         <div>
-          <button type="submit" className="btn waves-effect waves-light green">{displayName}</button>
+          <button type="submit" className="btn waves-effect waves-light green">
+            {displayName}
+          </button>
         </div>
         {error && error.response && <div> {error.response.data} </div>}
       </form>
-      <a href="/auth/google"><img className="google-btn" src="http://www.setyourowntests.com/_/rsrc/1468869481521/help/accounts/btn_google_signin_dark_normal_web%402x.png" /></a>
+      <a href="/auth/google">
+        <img
+          className="google-btn"
+          src="http://www.setyourowntests.com/_/rsrc/1468869481521/help/accounts/btn_google_signin_dark_normal_web%402x.png"
+        />
+      </a>
     </div>
-  )
-}
+  );
+};
 
 /**
  * CONTAINER
@@ -42,27 +52,27 @@ const AuthForm = (props) => {
  *   function, and share the same Component. This is a good example of how we
  *   can stay DRY with interfaces that are very similar to each other!
  */
-const mapLogin = (state) => {
+const mapLogin = state => {
   return {
     name: 'login',
     displayName: 'Login',
     error: state.user.error
-  }
-}
+  };
+};
 
-const mapDispatch = (dispatch) => {
+const mapDispatch = dispatch => {
   return {
-    handleSubmit (evt) {
-      evt.preventDefault()
-      const formName = evt.target.name
-      const email = evt.target.email.value
-      const password = evt.target.password.value
-      dispatch(auth(email, password, formName))
+    handleSubmit(evt) {
+      evt.preventDefault();
+      const formName = evt.target.name;
+      const email = evt.target.email.value;
+      const password = evt.target.password.value;
+      dispatch(auth(email, password, formName));
     }
-  }
-}
+  };
+};
 
-export default connect(mapLogin, mapDispatch)(AuthForm)
+export default connect(mapLogin, mapDispatch)(AuthForm);
 
 /**
  * PROP TYPES
@@ -72,4 +82,4 @@ AuthForm.propTypes = {
   displayName: PropTypes.string.isRequired,
   handleSubmit: PropTypes.func.isRequired,
   error: PropTypes.object
-}
+};
