+++ client/components/singleHome/home-content.js
