+++ client/components/singleHome/home-info.js
